v1.3
Decided to separate the Die charts to better navigation of the project.
No need to send strategy as parameter anymore. Changed Chart do DieChart
because there might be a PhaseChart in the future. We'll see.

Separated Phase 5 Corruption and Military in 2 different methods in the
Main. They might become separate classes in the future.